DeleteParser fix and test

+++ src/main/java/brokeculator/parser/DeleteParser.java
@@ -1,18 +1,24 @@
 package brokeculator.parser;
 
+import brokeculator.command.Command;
 import brokeculator.command.DeleteCommand;
+import brokeculator.command.InvalidCommand;
 
 public class DeleteParser {
-    public static DeleteCommand parseInput(String userInput) {
+    public static Command parseInput(String userInput) {
         String[] userInputAsArray = userInput.trim().split("\\s+");
         int indexToDelete = 0;
         if (userInputAsArray.length == 1) {
-            //TODO proper error handling when delete command has no index
-            return new DeleteCommand(0);
-        } else {
-            indexToDelete = Integer.parseInt(userInputAsArray[1]);
+            return new InvalidCommand("Delete index not specified");
         }
-
+        try {
+            indexToDelete = Integer.parseInt(userInputAsArray[1]);
+        } catch (NumberFormatException e) {
+            return new InvalidCommand("Delete index cannot be non-integer");
+        }
+        if (indexToDelete < 0) {
+            return new InvalidCommand("Delete index cannot be negative integer");
+        }
         return new DeleteCommand(indexToDelete);
     }
 }
